Refactoring Step 8a Introducing the State Pattern
I have split the introduction of the state pattern into three steps. The Steps are called Step a:"Replace Type Code with State/Strategy", Step b:"Move Method" and Step c:"Replace Conditional with Polymorphism". The introduction of the state pattern is required to replace the switch statement of the getCharge() Method of the "Movie" class.